requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Bcrypt==0.6.2
Flask-Login==0.2.11
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.11
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
SQLAlchemy==0.9.8
WTForms==2.0.2
Werkzeug==0.10.1
alembic==0.7.4
argparse==1.2.1
itsdangerous==0.24
oursql==0.9.3.1
python-bcrypt==0.3.1
requests==2.5.3
webassets==0.10.1
wsgiref==0.1.2
